projects/malaysia-fuel-dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyjwt==2.8.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Scraping & Data Ingestion
beautifulsoup4==4.12.2
requests==2.31.0
feedparser==6.0.10
PyPDF2==4.0.1
pytesseract==0.3.10

# Caching & Background Tasks
redis==5.0.1
celery==5.3.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring & Logging
python-json-logger==2.0.7
